Move/add cd at the start of every public function
There are two different programs in this same. One for managing tasks and one for managing logs. I don't want to split them in two by now for minimalism sake but also I don't want to treat one as 1st class citizen while the other is treated as 2nd class... It is a bit redundant but solve the problem and hopefully make the code easier to trace.